move position_type
