+++ packages/flame/lib/src/components/core/position_type.dart
